adding villa-booking-user types in types directory
