+++ web/src/types/i18.types.ts
@@ -1,9 +1,7 @@
 import type { supportedLanguages } from '@/utils/constants';
 
-// Supported language codes
 export type SupportedLanguage = typeof supportedLanguages[number];
 
-// Language configuration
 export interface Language {
      code: SupportedLanguage;
      name: string;
