+++ web/src/types/user.types.ts
@@ -4,6 +4,7 @@ export interface User {
     fullName: string;
     phone?: string;
     dateOfBirth?: Date;
+    isActive: boolean;
     role: 'GUEST' | 'HOST' | 'ADMIN';
     createdAt: string;
     updatedAt: string;
